Refactor: Extract StarRating component to eliminate code duplication
- Created reusable StarRating component in src/components/StarRating.jsx
- Removed duplicate Stars component from HomePage.jsx (37 lines)
- Removed duplicate Stars component from ProductPage.jsx (35 lines)
- Added configurable color prop to support both design variants
- Added JSDoc documentation for better developer experience
- Maintained exact same visual appearance and behavior

Benefits:
- Eliminates ~72 lines of duplicate code
- Single source of truth for star rating display
- Easier maintenance and future enhancements
- Better code organization following DRY principle
