+++ client/src/pages/HomePage.jsx
@@ -11,6 +11,7 @@ import FitnessChatBot from "../components/FitnessChatBot";
 import WelcomeBanner from "../components/WelcomeBanner";
 import { useWelcomeDiscount } from "../auth/useWelcomeDiscount";
 import BMICalculator from "../components/BMICalculator";
+import StarRating from "../components/StarRating";
 
 const API = import.meta.env.VITE_API_URL || "http://localhost:5000";
 
@@ -27,44 +28,7 @@ const PLANS = [
   { name: "Mobility & Recovery", duration: "8 Weeks", desc: "Flexibility-first programming, ideal for desk workers", tag: null, route: "/plans/mobility-recovery" },
 ];
 
-const Stars = ({ rating = 0, size = "sm" }) => {
-  const full = Math.floor(rating || 0);
-  const half = (rating || 0) % 1 >= 0.5 ? 1 : 0;
-  const empty = 5 - full - half;
-  const starPath = "M12 .587l3.668 7.431L24 9.748l-6 5.847L19.335 24 12 19.897 4.665 24 6 15.595 0 9.748l8.332-1.73L12 .587z";
-  const sizeClass = size === "lg" ? "w-4 h-4" : "w-3 h-3";
 
-  return (
-    <span className={`inline-flex items-center gap-0.5 ${size === "lg" ? "text-base" : "text-xs"}`} aria-hidden>
-      {Array.from({ length: full }).map((_, i) => (
-        <svg key={`full-${i}`} className={`${sizeClass} text-stone-500`} viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-          <path d={starPath} fill="currentColor" />
-        </svg>
-      ))}
-
-      {half ? (() => {
-        const id = `half-${Math.random().toString(36).slice(2)}`;
-        return (
-          <svg key="half" className={`${sizeClass} text-stone-500`} viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-            <defs>
-              <clipPath id={id}><rect x="0" y="0" width="12" height="24" /></clipPath>
-            </defs>
-            {/* faint empty star behind */}
-            <path d={starPath} fill="currentColor" className="text-stone-300" style={{ fill: 'currentColor', opacity: 0.28 }} />
-            {/* filled half */}
-            <path d={starPath} fill="currentColor" clipPath={`url(#${id})`} />
-          </svg>
-        );
-      })() : null}
-
-      {Array.from({ length: empty }).map((_, i) => (
-        <svg key={`empty-${i}`} className={`${sizeClass} text-stone-300`} viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-          <path d={starPath} fill="currentColor" style={{ opacity: 0.28 }} />
-        </svg>
-      ))}
-    </span>
-  );
-};
 
 function mapCart(cartDoc, products) {
   return cartDoc.items.map(it => {
@@ -149,7 +113,7 @@ function ProductCard({ product, onAdd, cartItems = [], updateQty }) {
         </h3>
 
         <div className="flex items-center gap-1 sm:gap-1.5 mb-2 sm:mb-3">
-          <Stars rating={product.rating} />
+          <StarRating rating={product.rating} />
           <span className="text-[9px] sm:text-[10px] text-stone-400">({product.reviews})</span>
         </div>
